+++ frontend/src/api_data/creditsIllustrationData.ts
