+++ frontend/src/components/components.tsx
@@ -9,8 +9,6 @@ import { Checkbox } from "./Inputs/Checkbox";
 import { Chip } from "./Inputs/Chip";
 import { Dropdown, DropdownOption } from "./Inputs/Dropdown";
 import { TextField } from "./Inputs/Textfield";
-import { HeaderNav } from "./Navigation/HeaderNav";
-import { SimpleNav } from "./Navigation/SimpleNav";
 import { FooterNav } from "./Navigation/FooterNav";
 import { ProgressBar } from "./Navigation/ProgressBar";
 import { Notification } from "./Notification/Notification";
@@ -37,8 +35,6 @@ export {
   DropdownOption,
   TextField,
   // Navigation
-  HeaderNav,
-  SimpleNav,
   FooterNav,
   ProgressBar,
   // Notification
