+++ frontend/src/layouts/HomeLayout.tsx
@@ -8,4 +8,5 @@ const HomeLayout = () => {
     </>
   );
 };
+
 export default HomeLayout;
