+++ frontend/src/tw-components/index.tsx
@@ -1,0 +1,3 @@
+export { default as AuthNav } from "./AuthNav";
+export { default as HeaderNav } from "./HeaderNav";
+export { default as InputGroup } from "./InputGroup";
