+++ frontend/tailwind.config.js
@@ -3,8 +3,6 @@ module.exports = {
   content: [
     "./src/pages/Demo/DemoTailwind.tsx",
     "./src/pages/Authentication/*.tsx",
-    "./src/components/Navigation/HeaderNav.tsx",
-    "./src/components/Navigation/SimpleNav.tsx",
     "./src/tw-components/*.tsx",
   ], // Will change to "./src/**/*.{js,jsx,tsx}", "./templates/index.html"
   theme: {
